Extract AnnotationDrawing component
